Edited bugs for SalaryLoand Schedule Entry

+++ Common/PresentationLayer/Models/DepartmentModel.vb
@@ -6,7 +6,7 @@
         Public Property DepartmentCode() As String
         Public Property DepartmentName() As String
         Public Property DepartmentNameAra() As String
-        Public Property ParentIdNo() As Integer?
+        Public Property ParentIdNo() As Int16?
         Public Property Notes() As String
         Public Property RevCostCenterIdNo() As Int16
         Public Property SortKey As String
